CC/requirements.txt
Pillow==11.0.0
flask==3.1.0
Numpy==2.0.2
tensorflow==2.18.0
firebase_admin==6.6.0